File Read(Updated)
Now can accept text files with a single line of coordinates and player designations. Returns a list of tuples called Coordinates, where each tuple is a particular chess piece's coordinates. 

Coordinate[0] is the tuple for the attacking player's King
Coordinate[1] is the tuple for the attacking player's Rook
Coordinate[2] is the tuple for the defending player's King

If the Coordinate List contains more than 3 tuples, the text file has too many coordinates.
The text file should be in identical format to that which is shown in the Assignment 1 sheet.